Move the golang files to gitoperations directory.
This change better supports golang module/package naming.  The suffix of the
module name "golang-gitoperations" in "github.com/amzn/golang-gitoperations" is
not compatible as a package name due to the hyphen.

The options are to choose a different repository name or place the .go
files in a subdirectory. I choose the latter approach.